Fix a stacked borrows violation in MIRI
This commit fixes an error which comes about when running `cargo
+nightly miri test` on a crate that depends on this library. MIRI
implements experimental and upcoming versions of aliasing models for
Rust to help assist with their refinement and clarification over time.
The "error" here isn't really a deal-breaker per se as it's a pretty
minor code change which I think is largely "just" appeasing MIRI.
Despite this though I figured I'd try to get this upstream as it'd help
us as we depend on this crate and are looking to run MIRI on our own
code soon.

The issue at hand here is pretty subtle, but I believe the idea is that
by calling `slice.as_mut_ptr()` a second time it invalidates the first
pointer returned. Instead the fix is to base an end pointer on the
starting pointer so both have the same provenance and are derived from
the same original pointer.